Fix data type select bandit

+++ gymnax/environments/misc/bernoulli_bandit.py
@@ -22,7 +22,7 @@ class EnvState:
 class EnvParams:
     reward_prob: float = 0.1
     normalize_time: bool = True
-    max_steps_in_episode: int = 100
+    max_steps_in_episode: float = 100.0
     min_lim: float = -1.0
     max_lim: float = 1.0
     t_max: int = 100
